core: move blockOffset getter closer to BlockInfra
