+++ core/kt-osrd-sim-infra/src/main/kotlin/fr/sncf/osrd/sim_infra/api/TrackInfra.kt
@@ -3,7 +3,9 @@ package fr.sncf.osrd.sim_infra.api
 import fr.sncf.osrd.reporting.exceptions.OSRDError.newUnknownTrackSectionError
 import fr.sncf.osrd.utils.Direction
 import fr.sncf.osrd.utils.indexing.*
+import fr.sncf.osrd.utils.units.DirOffset
 import fr.sncf.osrd.utils.units.Length
+import fr.sncf.osrd.utils.units.Offset
 
 /** Detectors detect when trains arrive and leave a given point location */
 sealed interface Detector
@@ -35,6 +37,10 @@ typealias DirDetectorId = DirStaticIdx<Detector>
 typealias DirTrackChunkId = DirStaticIdx<TrackChunk>
 
 typealias OptDirTrackChunkId = OptDirStaticIdx<TrackChunk>
+
+data class DirChunkLocation(val dirChunk: DirTrackChunkId, val offset: DirOffset<TrackChunk>)
+
+data class ChunkLocation(val chunk: TrackChunkId, val offset: Offset<TrackChunk>)
 
 fun <T> StaticIdx<T>.withDirection(dir: Direction): DirStaticIdx<T> {
     return DirStaticIdx(this, dir)
